requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.5.0
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.15.2
dotenv==0.9.9
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
